+++ src/TitleDetails/DeletePieceModal/index.js
@@ -1,0 +1,1 @@
+export { DeletePieceModal } from './DeletePieceModal';