+++ src/common/hooks/index.js
@@ -1,3 +1,5 @@
 export * from './usePieceMutator';
 export * from './useReceive';
 export * from './useQuickReceive';
+export * from './usePieces';
+export * from './useHoldingItems';