+++ src/common/hooks/usePieceMutator.js
@@ -9,22 +9,25 @@ import {
 
 import { getDehydratedPiece } from '../utils';
 
-export const usePieceMutator = (options = {}) => {
+export const usePieceMutator = (mutatorOptions = {}) => {
   const ky = useOkapiKy();
 
   const { mutateAsync } = useMutation({
-    mutationFn: (pieceValues) => {
+    mutationFn: ({ piece: pieceValues, options = {} }) => {
       const piece = getDehydratedPiece(pieceValues);
 
       const kyMethod = piece.id ? 'put' : 'post';
       const kyPath = piece.id ? `${ORDER_PIECES_API}/${piece.id}` : ORDER_PIECES_API;
-
-      return ky[kyMethod](kyPath, {
+      const kyOptions = {
+        method: options?.method ?? kyMethod,
         json: piece,
         searchParams: pieceValues.isCreateItem ? { createItem: true } : undefined,
-      }).then(response => (piece.id ? piece : response.json()));
+        ...options,
+      };
+
+      return ky(kyPath, kyOptions).then(response => (piece.id ? piece : response.json()));
     },
-    ...options,
+    ...mutatorOptions,
   });
 
   return {